+++ CHANGES.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added magnitude modifier. This is useful to determine the strength of an input vector, e.g. to determine how far a joystick is pushed. 
+
 ## [0.7.5] - 2025-08-19
 ### Fixed
 - Popup menus should now appear at the correct positions in multi-monitor and wide-screen setups ([#88](https://github.com/godotneers/G.U.I.D.E/issues/88), [#108](https://github.com/godotneers/G.U.I.D.E/issues/108)).